app/requirements.txt
click==7.1.2
fastapi==0.63.0
h11==0.12.0
pydantic==1.8.1
pyhumps==1.6.1
starlette==0.13.6
typing-extensions==3.7.4.3
uvicorn==0.13.4
sqlalchemy
fastapi-sqlalchemy
pony==0.7.14
requests
psycopg2-binary
yoyo-migrations==7.3.1
typer[all]
pytest==5.3.2
PyJWT==2.1.0
passlib==1.7.4
python-multipart==0.0.5
bcrypt==3.2.0
coverage